requirements.txt
requests>=2.28.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
streamlit>=1.0.0
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
pillow>=8.0.0
ollama>=0.1.0